Updated timing functions, added test cases to cross compiled main

+++ include/patlib.h
@@ -3,7 +3,7 @@
 #include <sys/time.h>
 
 long start;
-int msec;
-#define timing(func,name) start = clock(); func; msec = (clock() - start) * 1000 / CLOCKS_PER_SEC; printf("%s Timing: %d ms \n",name, msec);
+long msec;
+#define timing(func,name) start = clock(); func; msec = (clock() - start) * 1000 / CLOCKS_PER_SEC; printf("%s Timing: %lu ms \n",name, msec);
